Issue22

+++ churchSongs.js
@@ -47,7 +47,7 @@ function initSiteUI() {
   
     // flesh out any missing fields for clarity and simplicity of code
     healSongLibrary();
-  
+
     renderSongSetDropdown(
       'selNavSongSets', 
       'txtNavSongSetFilter', 
@@ -69,6 +69,7 @@ function initSiteUI() {
     // Nav formatting init
     ge('chkAllCaps').checked = songLibrary.defaults.allCaps ? 'checked' : '';
     ge('chkGenerateTitles').checked = songLibrary.defaults.generateTitle ? 'checked' : '';
+    renderNavSection();
   
     // song editing init
     hide('spnVerseUpdatedNotice');
@@ -78,6 +79,25 @@ function initSiteUI() {
   
     // AR UI init
     renderAspectRatioText();
+
+    checkForEmptySongsOrSongSets();
+}
+
+function checkForEmptySongsOrSongSets() {
+  const songSetName = ge('selNavSongSets').value;
+  const fIsSongSetMode = ge('chkNavSongSetMode').checked;
+  const fIsEmptySongSetSelected = 
+    fIsSongSetMode && 
+    songSetName && 
+    (!songLibrary.oSongSets[songSetName] ||
+      songLibrary.oSongSets[songSetName].length == 0);
+  const fHasSongs = Object.keys(songLibrary.oSongs).length > 0 && !fIsEmptySongSetSelected;
+  const cSongSets = Object.keys(songLibrary.oSongSets).length;
+  show('tdProjectorSection', fHasSongs);
+  show('trSongOrSongSetSelection', fHasSongs);
+  show('trNavSection', fHasSongs);
+  show('trGeneralFormattingSection', fHasSongs);
+  show('trSearch', fHasSongs || cSongSets > 0);
 }
 
 // projector
@@ -113,7 +133,7 @@ function getNavState() {
     nav.aSongsInSet = aSongsInSet;
     nav.iSongInSet = g.nav.iSongInSet;
     console.assert(nav.iSongInSet >= 0);
-    console.assert(nav.iSongInSet < nav.cSongsInSet);
+    // console.assert(nav.iSongInSet < nav.cSongsInSet);
     nav.songName = songLibrary.oSongSets[nav.songSetName][nav.iSongInSet];
     nav = addNavSongState(nav);
 
@@ -162,18 +182,20 @@ function addNavSongState(nav) {
   nav.iPageInSong = g.nav.iPageInSong;
   console.assert(nav.iPageInSong >= 0);
   let aPagesInSong = [];
-  console.assert(Number.isInteger(nav.songData.RepeatCount));
-  console.assert(nav.songData.RepeatCount > 0);
-  for (let i = 0; i < nav.songData.RepeatCount; i++) {
-    aPagesInSong = aPagesInSong.concat(nav.songData.aPageOrder);
-  }
-  if (nav.songData.TagPage) {
-    aPagesInSong = aPagesInSong.concat([ nav.songData.TagPage ]);
+  if (nav.songData) {
+    console.assert(Number.isInteger(nav.songData.RepeatCount));
+    console.assert(nav.songData.RepeatCount > 0);
+    for (let i = 0; i < nav.songData.RepeatCount; i++) {
+      aPagesInSong = aPagesInSong.concat(nav.songData.aPageOrder);
+    }
+    if (nav.songData.TagPage) {
+      aPagesInSong = aPagesInSong.concat([ nav.songData.TagPage ]);
+    }
   }
   nav.cPagesInSong = aPagesInSong.length;
   nav.pageName = aPagesInSong[nav.iPageInSong];
   nav.aPagesInSong = aPagesInSong;
-  console.assert(nav.iPageInSong < nav.cPagesInSong);
+  // console.assert(nav.iPageInSong < nav.cPagesInSong);
   nav.fInReview = false;
   nav.fBlankScreen = g.nav.fBlankScreen;
   nav.showTitlePage = g.nav.showTitlePage;
@@ -424,7 +446,7 @@ function renderSelectControl(
   aOptionValues=aOptionTexts) {
 
   let htmlOptions = '';
-  if (aOptionTexts) {
+  if (aOptionTexts && aOptionTexts.length) {
     aOptionTexts.forEach(
       (text, idx) => {
         htmlOptions += `<option value="${
@@ -438,7 +460,7 @@ function renderSelectControl(
     )
   }
   ge(idSel).innerHTML = htmlOptions;
-  ge(idSel).value = selectedValue;
+  ge(idSel).value = htmlOptions ? selectedValue : null;
 }
 
 function enableNavButtons() {
@@ -449,15 +471,18 @@ function enableNavButtons() {
     enableElement('btnPrevReviewPage', nav.iPageInReview > 0);
     enableElement('btnNextReviewPage', nav.iPageInReview < nav.cPagesInReview - 1);
   } else {
-    enableElement('btnNavPrevSong', nav.iSongInSet > 0);
-    enableElement('btnNextSong', nav.iSongInSet < nav.cSongsInSet - 1);
-    enableElement('btnNavPrevSongPage', !nav.fBlankScreen);
+    const fSongsInSet = nav.cSongsInSet > 0;
+    enableElement('btnNavPrevSong', fSongsInSet && nav.iSongInSet > 0);
+    enableElement('btnNextSong', fSongsInSet && nav.iSongInSet < nav.cSongsInSet - 1);
+    enableElement('btnNavPrevSongPage', fSongsInSet && !nav.fBlankScreen);
     enableElement('btnNavNextSongPage', 
-      nav.iPageInSong < nav.cPagesInSong - 1 || 
-      nav.iSongInSet < nav.cSongsInSet - 1 ||
-      (nav.iPageInSong == 0 && // special case for one page song
-        nav.cPagesInSong == 1 && 
-        (nav.fBlankScreen || nav.showTitlePage)
+      fSongsInSet && (
+        nav.iPageInSong < nav.cPagesInSong - 1 || 
+        nav.iSongInSet < nav.cSongsInSet - 1 ||
+        (nav.iPageInSong == 0 && // special case for one page song
+          nav.cPagesInSong == 1 && 
+          (nav.fBlankScreen || nav.showTitlePage)
+        )
       )
     );
   }
@@ -510,12 +535,14 @@ function blankScreen(event) {
 }
 
 function renderNavSection() {
-  const oMsg = getMessageFromGlobals();
-  localStorage.setItem('projector-message', JSON.stringify(oMsg));
-  localStorage.clear();
-  renderNavPagePreview();
-  renderNavStateText();
-  enableNavButtons();
+  if (getNavState().songData) {
+    const oMsg = getMessageFromGlobals();
+    localStorage.setItem('projector-message', JSON.stringify(oMsg));
+    localStorage.clear();
+    renderNavPagePreview();
+    renderNavStateText();
+    enableNavButtons();
+  }
 }
 
 function renderNavPagePreview() {
@@ -978,7 +1005,7 @@ function onSelCurrentSongSetChanged() {
   enableSongSetEditButtons();
 }
 
-async function deleteSongSet(event) {
+function deleteSongSet(event) {
   const songSetToDelete = ge('selAllSongSetsToEdit').value;
   if (getAllSongSetNames().includes(songSetToDelete)) {
     delete songLibrary.oSongSets[songSetToDelete];
@@ -989,6 +1016,7 @@ async function deleteSongSet(event) {
       '',
       'spnNoSongSetsToEdit');
   }
+  checkForEmptySongsOrSongSets();
 }
 
 function createNewSongSet(event) {
@@ -1005,6 +1033,7 @@ function createNewSongSet(event) {
     ge('selNavSongSets').value,
     'spnNoSongSetsToEdit');
   editSelectedSongSet();
+  checkForEmptySongsOrSongSets();
 }
 
 function copySongSet(event) {
@@ -1104,11 +1133,13 @@ function addSongToNewSongSet(event) {
       sses.aSongList.shift();
     }
     songLibrary.oSongSets[sses.songSetNameToEdit] = sses.aSongList;
-    getSongSetEditState();
     renderSelectControl(
       'selCurrentSongSet',
       sses.aSongList, 
       sses.songNameToAdd);
+    fillSongToEdit();
+    enableSongSetEditButtons();
+    checkForEmptySongsOrSongSets();
   }
 }
 
@@ -1126,7 +1157,7 @@ async function moveSongUpInSongSet(event) {
     T);
 }
 
-async function moveSongDownInSongSet(event) {
+function moveSongDownInSongSet(event) {
   const sses = getSongSetEditState();
   console.assert(sses.iSelectedSongInSet < sses.aSongList.length - 1);
   const T = sses.aSongsInSongSet[sses.iSelectedSongInSet];
@@ -1140,13 +1171,14 @@ async function moveSongDownInSongSet(event) {
     T);
 }
 
-async function deleteSongFromSongSet(event) {
+function deleteSongFromSongSet(event) {
   const sses = getSongSetEditState();
   sses.aSongsInSongSet.splice(sses.iSelectedSongInSet, 1);
   renderSelectControl(
     'selCurrentSongSet', 
     sses.aSongsInSongSet);
   editSelectedSongSet();
+  checkForEmptySongsOrSongSets();
 }
 
 // Edit or Create a Song
@@ -1167,23 +1199,23 @@ function fillSongToEdit(event) {
     ses.aPageNames, 
     ge('selSongVersesToEdit').value);
   onSelectedVerseToEditChanged(null, ses);
-  onVerseOrderChanged();
 
+  // other settings
   if (ses.songData) {
-    // other settings
+    const sd = ses.songData;
     ge('txtNotes').value = 
-      ses.songData.Notes ? ses.songData.Notes : '';
+      sd.Notes ? sd.Notes : '';
     ge('txtTitleNote').value = 
-      ses.songData.TitleNote ? ses.songData.TitleNote : '';
+      sd.TitleNote ? sd.TitleNote : '';
     ge('txtAuthor').value = 
-      ses.songData.Author ? ses.songData.Author : '';
+      sd.Author ? sd.Author : '';
     ge('txtPublisher').value = 
-      ses.songData.Publisher ? ses.songData.Publisher : '';
+      sd.Publisher ? sd.Publisher : '';
     ge('txtLicense').value = 
-      ses.songData.License ? ses.songData.License : '';
+      sd.License ? sd.License : '';
     ge('txtDefaultLicense').value = songLibrary.defaults.License;
     ge('selRepeatCount').value = 
-      ses.songData.RepeatCount ? ses.songData.RepeatCount : 1;
+      sd.RepeatCount ? sd.RepeatCount : 1;
   } else {
     ge('txtNotes').value = '';
     ge('txtTitleNote').value = '';
@@ -1322,7 +1354,7 @@ function enableSongEditButtons() {
   enableSongEditVerseButtons(ses);
 }
 
-function createEmptySong(event) {
+function createNewSong(event) {
   const ses = getSongEditState();
   const sExistingSongName = 
     doesProposedEditSongNameExist(ses.newSongEditName);
@@ -1347,8 +1379,11 @@ function createEmptySong(event) {
   ge('txtNewSongName').value = '';
   onNewSongFilterChanged();
   ge('selAllSongsToEdit').value = ses.newSongEditName;
+  initSongValues(ses.newSongEditName);
   fillSongToEdit();
   ge('txtaVerseLines').value = '';
+  enableSongSetEditButtons();
+  checkForEmptySongsOrSongSets();
 }
 
 function clearSongUI(event) {
@@ -1431,7 +1466,7 @@ function reRenderAllSongSelectControls(oldSongName, newSongName) {
     'spnNoSongsToEdit');
 }
 
-async function deleteSong(event) {
+function deleteSong(event) {
   const ses = getSongEditState();
   delete songLibrary.oSongs[ses.selectedSongToEdit];
 
@@ -1452,6 +1487,8 @@ async function deleteSong(event) {
   // this makes the song dropdowns select the first song.
   reRenderAllSongSelectControls();
   initSongEditUI();
+  enableSongSetEditButtons();
+  checkForEmptySongsOrSongSets();
 }
  
 // edit song verses
@@ -1511,8 +1548,7 @@ function onSelectedVerseToEditChanged(event=null, ses) {
         '', 
       Object.keys(ses.songData.aPageOrder)
       );
-    ge('chkIsTagVerse').checked = 
-      ses.songData.TagPage == ses.pageName ? 'checked' : '';
+    ge('chkIsTagVerse').checked = ses.fIsTagVerse ? 'checked' : '';
   }
   setSongVerseError('');
   renderOverallOrderText();
@@ -1542,7 +1578,7 @@ function onVerseLinesChanged(event) {
   show('spnVerseUpdatedNotice');
   setTimeout(() => {
     hide('spnVerseUpdatedNotice');
-  }, 1000);
+  }, 3000);
 }
 
 function onNewVerseNameChanged (event) {
@@ -2574,28 +2610,29 @@ function healSongLibrary() {
   // fix up the song oSongs
   Object.keys(o.oSongs).forEach(
     function (songName) {
-      const oSD = o.oSongs[songName];
-      const aPageNames = Object.keys(oSD.oPages);
-
-      // fix up song data values
-      if (!oSD.RepeatCount) oSD.RepeatCount = 1;
-      if (!oSD.TitleNote) oSD.TitleNote = '';
-      if (!oSD.Notes) oSD.Notes = '';
-      if (!oSD.Author) oSD.Author = '';
-      if (!oSD.Publisher) oSD.Publisher = '';
-      if (!oSD.License) oSD.License = songLibrary.defaults.License;
-      console.assert(o.defaults.fontSize);
-      if (!oSD.fontSize) oSD.fontSize = o.defaults.fontSize;
-      console.assert(o.defaults.fontBoldness);
-      if (!oSD.fontBoldness) oSD.fontBoldness = o.defaults.fontBoldness;
-      console.assert(o.defaults.lineHeight);
-      if (!oSD.lineHeight) oSD.lineHeight = o.defaults.lineHeight;
-      // oPageData was removed
-      if (oSD.oPageData) {
-        delete oSD.oPageData;
-      }
+      initSongValues(songName);
     }
   ); // oSongs 
+}
+
+function initSongValues(songName) {
+  const o = songLibrary;
+  const oSD = o.oSongs[songName];
+  const aPageNames = Object.keys(oSD.oPages);
+
+  // fix up song data values
+  if (!oSD.RepeatCount) oSD.RepeatCount = 1;
+  if (!oSD.TitleNote) oSD.TitleNote = '';
+  if (!oSD.Notes) oSD.Notes = '';
+  if (!oSD.Author) oSD.Author = '';
+  if (!oSD.Publisher) oSD.Publisher = '';
+  if (!oSD.License) oSD.License = songLibrary.defaults.License;
+  console.assert(o.defaults.fontSize);
+  if (!oSD.fontSize) oSD.fontSize = o.defaults.fontSize;
+  console.assert(o.defaults.fontBoldness);
+  if (!oSD.fontBoldness) oSD.fontBoldness = o.defaults.fontBoldness;
+  console.assert(o.defaults.lineHeight);
+  if (!oSD.lineHeight) oSD.lineHeight = o.defaults.lineHeight;
 }
 
 function getUnwoundPages(sd) {
